ecsite/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
gspread>=6.1.0
google-auth>=2.35.0

